established db connection to azure mysql
