+++ fmi-mds-app/src/main/java/com/mds/DataBaseConnection/DatabaseServices.java
@@ -1,0 +1,17 @@
+package com.mds.DataBaseConnection;
+
+public interface DatabaseServices {
+
+    String insertTable = "INSERT INTO `users`\n" +
+            "(`username`,\n" +
+            "`email`,\n" +
+            "`password`,\n" +
+            "`create_time`)\n" +
+            "VALUES\n" +
+            "(?,\n" +
+            "?,\n" +
+            "?,\n" +
+            "?);";
+    String deleteUser = "DELETE FROM `users` WHERE `username` = ?;";
+
+}
